refactor: profile 관련 기능을 profileService로 분리
