+++ src/main/java/Trendithon/SpinOff/domain/member/controller/MemberController.java
@@ -3,6 +3,7 @@ package Trendithon.SpinOff.domain.member.controller;
 import Trendithon.SpinOff.domain.member.dto.*;
 import Trendithon.SpinOff.domain.member.entity.Member;
 import Trendithon.SpinOff.domain.member.service.MemberService;
+import Trendithon.SpinOff.domain.member.service.ProfileService;
 import Trendithon.SpinOff.global.jwt.service.TokenService;
 import Trendithon.SpinOff.global.jwt.dto.TokenDto;
 import jakarta.validation.Valid;
@@ -24,11 +25,13 @@ import java.util.Optional;
 @RequestMapping("/api")
 public class MemberController {
     private final MemberService memberService;
+    private final ProfileService profileService;
     private final TokenService tokenService;
 
     @Autowired
-    public MemberController(MemberService memberService, TokenService tokenService) {
+    public MemberController(MemberService memberService, ProfileService profileService, TokenService tokenService) {
         this.memberService = memberService;
+        this.profileService = profileService;
         this.tokenService = tokenService;
     }
     
@@ -64,15 +67,15 @@ public class MemberController {
     @PostMapping("/information/add")
     public ResponseEntity<Boolean> addInformation(@RequestBody Information information) {
         log.info("멤버 아이디 = {}", information.getMemberId());
-        boolean resultInfo = memberService.addInformation(information);
-        boolean resultTechnic = memberService.addTechnic(information.getMemberId(), information.getTechnics());
+        boolean resultInfo = profileService.addInformation(information);
+        boolean resultTechnic = profileService.addTechnic(information.getMemberId(), information.getTechnics());
         return ResponseEntity.ok(resultInfo && resultTechnic);
     }
 
     @PostMapping("information/edit")
     public ResponseEntity<Boolean> editInformation(@RequestBody EditInformation editInformation) {
-        boolean result = memberService.editInformation(editInformation);
-        boolean resultTechnic = memberService.editTechnics(editInformation);
+        boolean result = profileService.editInformation(editInformation);
+        boolean resultTechnic = profileService.editTechnics(editInformation);
         return ResponseEntity.ok(result && resultTechnic);
     }
 }
