+++ src/main/java/Trendithon/SpinOff/domain/member/repository/ProfileTechnicJpaRepository.java
@@ -2,10 +2,7 @@ package Trendithon.SpinOff.domain.member.repository;
 
 import Trendithon.SpinOff.domain.member.entity.ProfileTechnic;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
-public interface ProfileTechnicJpaRepository extends JpaRepository<ProfileTechnic, Long> {
-    @Override
-    Optional<ProfileTechnic> findById(Long id);
-}
+@Repository
+public interface ProfileTechnicJpaRepository extends JpaRepository<ProfileTechnic, Long> {}
